refactor: split currency API module by responsibility
