+++ frontend/src/api/currency/index.ts
@@ -1,0 +1,3 @@
+export type { Currency } from '@/api/currency/types';
+export { fetchCurrencies } from '@/api/currency/requests';
+export { useCurrencies } from '@/api/currency/hooks';
