+++ frontend/src/api/currency/types.ts
@@ -1,0 +1,6 @@
+export interface Currency {
+  id: string;
+  name: string;
+  symbol: string;
+  minor_unit_exponent: number;
+}
